requirements.txt
flask==2.3.3
gunicorn==21.2.0
flask-cors==4.0.0
openai==1.3.7
python-dotenv==1.0.0